Move LibretroMenuStyle to raylib-libretro-styles.h
